From: chris.merrett melbourne.co.uk
I've been tasked with moving some of our Windows-based internal systems over
to *nix-based platforms, and the time has come to migrate from PRTG (spit)
to MRTG w/ RRDTool.  One of the main problems I encountered during my
initial research involved low-resolution data storage, and the fact that
MRTG doesn't output RRD's with user-definable row counts for
lower-resolution RRA's.  Setting 5-min row counts via cfgmaker was fine via
RRDRowCount[], but no such options exist for 30 minute, 2 hour and 1 day.

When setting an RRDRowCount of 17856 (62 days of 5 minute data), this some
what makes the default 30 minute and 2 hour RRA's redundant, as they'd be
storing much less data with just 800 rows.

As a result, I am submitting a patch for the main mrtg script and
MRTG_lib.pm to allow for three new config options to compliment the existing
RRDRowCount option.  These are:

RRDRowCount30m
RRDRowCount2h
RRDRowCount1d

These options are passed to cfgmaker with the --global parameter just as
before.  If not passed as options, they simply default to 800 so as to not
break MRTG's current behavior.  I'm not entirely sure if there is any reason
why these options have not been included in the past, but I certainly feel
there is justifiable reason for them to be considered for inclusion in the
next release of MRTG.