Fix chdir to workdir in CGI
This is a patch from Debian.

Original name: 040-bts113814_mrtgindx_cgi.patch
URL: https://salsa.debian.org/eriberto/mrtg/-/blob/debian/master/debian/patches/040-bts113814_mrtgindx_cgi.patch

The header from this patch is available below:

Description: fix a comment error causing doesn't chdir to workdir
Author: Ryan Murray rmurray@cyberhqz.com
Reviewed-By: Sandro Tosi morph@debian.org
Bug-Debian: https://bugs.debian.org/113814
Last-Update: 2008-06-03

An explanation from Debian bug #113814:

the mrtgindx.cgi attempts to stat the generated png image, to display the time
it was generated after the title.  This fails unless the cgi is in the same
dir as all of the mrtg output.  This isn't the case on all Debian machines
at least, and a fix would be to chdir to the WorkDir value when it is found in
the config file.

The att.ping line should be commented out, the config file location should
match the default, and the Count.cgi usage should be commented out by default
(to match the comment above it saying to uncomment it).